+++ src/configUtils.ts
@@ -1,11 +1,12 @@
 import { GraphQLExtensionDeclaration } from "graphql-config";
 import { directiveNodes } from "./relayDirectives";
 
-export const RelayExtension: GraphQLExtensionDeclaration = (api) => {
+export const RelayDirectivesExtension: GraphQLExtensionDeclaration = (api) => {
   api.loaders.schema.use((document) => ({
     ...document,
     definitions: [...document.definitions, ...directiveNodes],
   }));
+
   return {
     name: "VScodeReasonML",
   };