+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master
 
+- Automatically restart RLS (if it's installed) whenver the Relay compiler has changed generated files. This works around the issue where RLS does not pick up that the Relay compiler has emitted new files, or changed existing generated files.
+
 ## 0.2.1
 
 - Update README with a list of the current features of the extension.